requirements.txt
tqdm==4.43.0
plotly==4.5.4
pandas==1.0.3
numpy>=1.18.2
open3d==0.12.0
torch==1.7.0
torchvision>=0.8.0
tensorboard==2.2.1
git+https://github.com/kornia/kornia
Pillow==7.1.1
logzero==1.5.0
pytest==5.4.2
imageio==2.8.0
matplotlib==3.2.1
codetiming==1.2.0
PyGeometry>=1.5.6
pytorch3d>=0.2.5
# Development
black==19.10b0
flake8==3.7.9
isort==4.3.21
